Refactoring favorite.js
J'en doute mais j'y bosse c'est quand même mieux maitenant
