+++ .gitignore
@@ -43,6 +43,7 @@ logs
 
 # Build output / Generated files
 assets/css/styles.css
+OLD FAVORITES
 
 # Autres fichiers générés
 dist/
