+++ assets/js/components/favorites/favoritesindex.js
@@ -1,0 +1,2 @@
+// favoritesindex.js - Point d'entrée ES module pour le module favoris
+export { FavoritesUI } from './favorite-ui.js';
