+++ index.html
@@ -40,7 +40,7 @@
                         const resourcesToPrefetch = [
                             'assets/js/app.js',
                             'assets/js/components/visualizer.js',
-                            'assets/js/components/favorites.js'
+                            'assets/js/components/favorites/favoritesModule.js'
                         ];
                         
                         resourcesToPrefetch.forEach(url => {
@@ -730,13 +730,13 @@
     </script>
     <script type="module" src="assets/js/locale-switcher.js" defer></script>
     <script src="assets/js/utils/helpers.js" defer></script>
-    <script src="assets/js/utils/storage.js" defer></script>
+    <script type="module" src="assets/js/utils/storage.js" defer></script>
     <script src="assets/js/utils/preferences.js" defer></script>
     <script src="assets/js/utils/constraintHelpers.js" defer></script>
     <script src="assets/js/components/contextMenu.js" defer></script>
     <script src="assets/js/components/notifications.js" defer></script>
     <script src="assets/js/components/visualizer.js" defer></script>
-    <script src="assets/js/components/favorites.js" defer></script>
+    <script type="module" src="assets/js/components/favorites/favoritesModule.js" defer></script>
     <script src="assets/js/components/tabletSelector.js" defer></script>
     
     <script>
@@ -777,7 +777,7 @@
         }
     </script>
     
-    <script src="assets/js/app.js"></script>
+    <script type="module" src="assets/js/app.js"></script>
     
     <!-- Initial rectangle positioning script -->
     <script>
